# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cakra Prima",
        "description": "Dokumentasi API Kas & Keuangan Cakra Prima",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cash-flow-periods/{period}/transactions": {
            "post": {
                "tags": [
                    "Cash Flow Transactions"
                ],
                "summary": "Tambah Transaksi Arus Kas",
                "operationId": "8502b01edcf1bba66939342616341a02",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trans_date",
                                    "description",
                                    "source"
                                ],
                                "properties": {
                                    "trans_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-05"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Pembelian ATK"
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "CASH"
                                    },
                                    "out_amount": {
                                        "type": "number",
                                        "example": 50000
                                    },
                                    "in_amount": {
                                        "type": "number",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/api/cash-flow-periods/{period}/transactions/{transaction}": {
            "put": {
                "tags": [
                    "Cash Flow Transactions"
                ],
                "summary": "Update Transaksi Arus Kas",
                "operationId": "197823bae95c33d5a04358da1c84a7f9",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "trans_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "source": {
                                        "type": "string"
                                    },
                                    "out_amount": {
                                        "type": "number"
                                    },
                                    "in_amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cash Flow Transactions"
                ],
                "summary": "Hapus Transaksi Arus Kas",
                "operationId": "9f5e8daf91ddeea4e23b7640c1cb7496",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
        "/api/debts": {
            "get": {
                "tags": [
                    "Debts"
                ],
                "summary": "Daftar Hutang Piutang",
                "operationId": "347c69ab1e5592b389254d9bd10b6ad9",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Debts"
                ],
                "summary": "Tambah Hutang Piutang",
                "operationId": "6602577612cef1d0ecc2a08c6d6bcd6a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "person_name",
                                    "amount",
                                    "status",
                                    "trans_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "person_name": {
                                        "type": "string",
                                        "example": "Mas Ryan"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 5000000
                                    },
                                    "paid_amount": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "belum lunas"
                                    },
                                    "trans_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-27"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/api/debts/{id}": {
            "get": {
                "tags": [
                    "Debts"
                ],
                "summary": "Detail Hutang Piutang",
                "operationId": "af50ce7e611e06a5f0886b246e87f247",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "put": {
                "tags": [
                    "Debts"
                ],
                "summary": "Update Hutang Piutang",
                "operationId": "8a524a80413a48645f8cea4bdbae5c64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "person_name": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "paid_amount": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "trans_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Debts"
                ],
                "summary": "Hapus Hutang Piutang",
                "operationId": "efda5885b17af25d67d337373236aeec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
...
    },
...
}